+++ src/RobinhoodCli/Commands/AccountCommand.cs
@@ -1,6 +1,8 @@
 using System.Linq;
 using System.Threading.Tasks;
 using Deadlock.Robinhood;
+using RobinhoodCli.ExecutionResults;
+using RobinhoodCli.Models;
 
 namespace RobinhoodCli.Commands
 {
