+++ src/RobinhoodCli/Commands/OrderCommand.cs
@@ -3,6 +3,8 @@ using System.Linq;
 using System.Threading.Tasks;
 using Deadlock.Robinhood.Model;
 using Deadlock.Robinhood;
+using RobinhoodCli.ExecutionResults;
+using RobinhoodCli.Models;
 
 namespace RobinhoodCli.Commands
 {
