+++ src/RobinhoodCli/Commands/PositionsCommand.cs
@@ -1,6 +1,8 @@
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using Deadlock.Robinhood;
+using RobinhoodCli.ExecutionResults;
+using RobinhoodCli.Models;
 
 namespace RobinhoodCli.Commands
 {
