+++ src/RobinhoodCli/Models/ExecutionContext.cs
@@ -2,7 +2,7 @@ using System.Collections.Generic;
 using Deadlock.Robinhood.Model;
 using RobinhoodCli.Commands;
 
-namespace RobinhoodCli
+namespace RobinhoodCli.Models
 {
     public class ExecutionContext
     {
@@ -10,11 +10,11 @@ namespace RobinhoodCli
         public Account ActiveAccount { get; set; }
         public List<OpenPosition> OpenPositions { get; set; }
 
-        public Queue<Commands.ICommand> CommandQueue { get; private set; }
+        public Queue<ICommand> CommandQueue { get; private set; }
 
         public ExecutionContext()
         {
-            CommandQueue = new Queue<Commands.ICommand>();
+            CommandQueue = new Queue<ICommand>();
         }
     }
 }
